# server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MetricFlow API",
    "version": "1.0.0",
    "description": "API documentation for MetricFlow Backend"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local Development Server"
    },
    {
      "url": "/api",
      "description": "Relative path (Production)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/admin/transfers": {
      "get": {
        "summary": "Get all transfers (queue)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "businessId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search by reference, recipient name/account, business name/email"
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter by start date (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter by end date (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of transfers"
          }
        }
      }
    },
...
    "/api/transfers/account-lookup": {
      "post": {
        "summary": "Lookup account details",
        "tags": [
          "Transfers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "bank_code",
                  "account_number"
                ],
                "properties": {
                  "bank_code": {
                    "type": "string"
                  },
                  "account_number": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account details"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/transfers/bulk": {
      "post": {
        "summary": "Initiate bulk transfer",
        "tags": [
          "Transfers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "transfers"
                ],
                "properties": {
                  "transfers": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "recipient_account",
                        "recipient_bank",
                        "recipient_name",
                        "amount"
                      ],
                      "properties": {
                        "recipient_account": {
                          "type": "string"
                        },
                        "recipient_bank": {
                          "type": "string"
                        },
                        "recipient_name": {
                          "type": "string"
                        },
                        "amount": {
                          "type": "number"
                        },
                        "remark": {
                          "type": "string"
                        },
                        "source_type": {
                          "type": "string"
                        },
                        "source_id": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transfers queued"
          }
        }
      }
    },
...
    "/api/transfers/{id}/retry": {
      "post": {
        "summary": "Retry a failed transfer",
        "tags": [
          "Transfers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transfer retried"
          }
        }
      }
    },
...
    "/transfers/bulk": {
      "post": {
        "summary": "Initiate a bulk transfer",
        "tags": [
          "Transfers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "type",
                  "otp"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "manual",
                      "sprint",
                      "salary",
                      "task"
                    ]
                  },
                  "otp": {
                    "type": "string",
                    "description": "OTP code for authorization"
                  },
                  "source_wallet_id": {
                    "type": "string",
                    "description": "ID of the wallet to fund the transfer from"
                  },
                  "data": {
                    "type": "object",
                    "description": "Data depending on type (manual items, sprint name, task IDs)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transfers queued"
          }
        }
      }
    },
    "/transfers": {
      "get": {
        "summary": "Get transfer queue",
        "tags": [
          "Transfers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search by recipient name, account or reference"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "processing",
                "success",
                "failed"
              ]
            },
            "description": "Filter by status"
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter by start date (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter by end date (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of transfers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/transfers/{id}/retry": {
      "post": {
        "summary": "Retry a failed transfer",
        "tags": [
          "Transfers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retry initiated"
          }
        }
      }
    },
...
  },
...
}